generative-ai/coursera-references/langfuse-evaluation-main/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# AI Providers
google-generativeai>=0.3.0  # For Gemini API
google-cloud-aiplatform>=1.36.0  # For Vertex AI (optional)

# Utilities
python-dotenv>=1.0.0  # For environment variables (optional)
tqdm>=4.65.0  # Progress bars (optional)
